Move calc2df to calc_ubi
